requirements.txt
fastapi
uvicorn[standard]
Pillow
torch
torchvision